Add function to get rendered canvas as string
Previously, using the API it was only possible to store the rendered
canvas into a file or output it in stdout. When rendering on
server-side, however, it is impractical to first write the output to a
file, only to read it again later.

This commit adds the function canvas.render_str() to get the rendered
image as string.